# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "نبض الحياة API",
        "description": "API للتطبيق الخيري نبض الحياة",
        "contact": {
            "email": "info@nabdhayah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "الحصول على الملف الشخصي",
                "operationId": "1a4b846fd60226473a8cb32221eea509",
                "responses": {
                    "200": {
                        "description": "الملف الشخصي",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "تحديث الملف الشخصي",
                "operationId": "45695aa3f758e4931d6236815e2a8cb6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "أحمد محمد"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "01234567890"
                                    },
                                    "current_password": {
                                        "type": "string",
                                        "example": "oldpassword"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "newpassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث الملف الشخصي بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم تحديث الملف الشخصي بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/image": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "تحديث صورة الملف الشخصي",
                "operationId": "0fbc35f0b14483f4ec9a2c8a3c3c2d74",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "image"
                                ],
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث الصورة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم تحديث الصورة بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "أحمد محمد"
                    },
                    "mobile": {
                        "type": "string",
                        "example": "01234567890"
                    },
                    "wallet_balance": {
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "profile_image": {
                        "type": "string",
                        "example": "http://example.com/image.jpg",
                        "nullable": true
                    },
                    "is_admin": {
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}